moved all api calls related to books into one file
