--- api/books/removeFromList.ts
@@ -1,18 +1,0 @@
-import { baseWithAuth } from "api/base"
-import { BookType } from "api/types"
-import { errorHandler } from "api/utils"
-
-export const removeFromList = async (id: number): Promise<BookType> => {
-  return await baseWithAuth()
-    .request({
-      url: `/api/user_books/${id}`,
-      method: "DELETE",
-    })
-    .then((res) => {
-      const book = res.data
-      return book
-    })
-    .catch((error) => {
-      errorHandler(error)
-    })
-}
